requirements.txt
fastapi
python-dotenv
retell-sdk
uvicorn
openai
